Resolve some static analyzer warnings

+++ src/decoder_strategy.h
@@ -53,11 +53,11 @@ public:
         read_cache_ = read_cache_ << length;
     }
 
-    static void on_line_begin(const size_t /*pixel_count*/, void* /*ptypeBuffer*/, int32_t /*pixelStride*/) noexcept
+    static void on_line_begin(const size_t /*pixel_count*/, void* /*source*/, size_t /*pixel_stride*/) noexcept
     {
     }
 
-    void on_line_end(const size_t pixel_count, const void* source, const int32_t pixel_stride) const
+    void on_line_end(const size_t pixel_count, const void* source, const size_t pixel_stride) const
     {
         process_line_->new_line_decoded(source, pixel_count, pixel_stride);
     }
